working on listing items

+++ shared/src/commonMain/kotlin/com/kaesik/tabletopwarhammer/library/data/library/dto/ClassDto.kt
@@ -16,5 +16,5 @@ data class ClassDto(
     @SerialName("careers")
     val careers: String,
     @SerialName("page")
-    val page: Double
+    val page: Int
 )
